// src/core/enum.ts
export enum CacheKeys {
  ALL_PAIRS = "ALL_PAIRS",
  ALL_PAIR_GRAPH = "ALL_PAIR_GRAPH",
  ORDERBOOKS = "ORDERBOOKS",
  ACCOUNT_INFO = "ACCOUNT_INFO",
  ROUTES = "ROUTES",
  AVAILABLE_PAIRS = "AVAILABLE_PAIRS",
  PLAN = "plan"
}

export enum Direction {
  BUY,
  SELL
}

export enum Exchange {
  poloniex = "poloniex",
  kraken = "kraken"
}